+++ src/6-Vault.sol
@@ -16,14 +16,19 @@ import "openzeppelin-contracts/contracts/token/ERC20/IERC20.sol";
  * Check your token balances using: forge script script/TokenStatus.s.sol --rpc-url $RPC_URL
  */
 contract Vault is Solvable {
-
     // Predefined list of ERC20 tokens
     IERC20[] public allowedTokens;
     mapping(address => mapping(IERC20 => uint256)) public balances;
     mapping(IERC20 => uint256) public totalBalances;
 
     constructor() {
-        allowedTokens = [IERC20(0x6B175474E89094C44Da98b954EedeAC495271d0F), IERC20(0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48), IERC20(0xD37EE7e4f452C6638c96536e68090De8cBcdb583), IERC20(0xA17581A9E3356d9A858b789D68B4d866e593aE94), IERC20(0xbe0Ed4138121EcFC5c0E56B40517da27E6c5226B)];
+        allowedTokens = [
+            IERC20(0x6B175474E89094C44Da98b954EedeAC495271d0F),
+            IERC20(0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48),
+            IERC20(0xD37EE7e4f452C6638c96536e68090De8cBcdb583),
+            IERC20(0xA17581A9E3356d9A858b789D68B4d866e593aE94),
+            IERC20(0xbe0Ed4138121EcFC5c0E56B40517da27E6c5226B)
+        ];
     }
 
     // Modifier to check if the token is allowed
@@ -32,14 +37,13 @@ contract Vault is Solvable {
         _;
     }
 
-
-    function allAllowedTokens() external view returns(IERC20[] memory){
+    function allAllowedTokens() external view returns (IERC20[] memory) {
         return allowedTokens;
     }
 
     // Function to check if a token is allowed
     function isTokenAllowed(IERC20 token) public view returns (bool) {
-        for (uint i = 0; i < allowedTokens.length; i++) {
+        for (uint256 i = 0; i < allowedTokens.length; i++) {
             if (allowedTokens[i] == token) {
                 return true;
             }
@@ -66,9 +70,9 @@ contract Vault is Solvable {
         require(token.transfer(msg.sender, amount), "Transfer failed");
     }
 
-    function isSolved() external view returns(bool) {
-        for (uint i = 0; i < allowedTokens.length; i++) {
-            if (totalBalances[allowedTokens[i]] >= 2**128) {
+    function isSolved() external view returns (bool) {
+        for (uint256 i = 0; i < allowedTokens.length; i++) {
+            if (totalBalances[allowedTokens[i]] >= 2 ** 128) {
                 return true;
             }
         }
